refactor: payment 타입 및 api 호출 응집
