+++ virtualFittingApp/src/features/initiate-checkout-single/index.ts
@@ -1,3 +1,2 @@
-export * from "./api/payment.api";
 export * from "./ui/initiate-checkout-single-button";
 export * from "./hooks/use-initiate-checkout";
